Enabled Git support for the winery-repository
In FilebasedRepository.java, determineRepositoryPath became protected so it can be used in derived classes.
It is used in GitBasedRepository.java to get the path of the local winery-repository.

The old GitBasedRepository.java is updated so that it doesn't use a remote anymore.
Thus the settings for user and password on the remote were removed.
Every save action is committed. The committing is blocking.

Renamed addCommitPush to addCommit since there is no pushing anymore.
All references to the method were updated.

Signed-off-by: Armin Hueneburg <hueneburg.armin@gmail.com>