Added backend support for Git in Artifacts
ImplementationArtifacts can now have an associated Git repo that provides the files.
The repo is cloned and traversed.
The URL and branch are specified with the namespace "http://www.opentosca.org/winery/extensions/tosca/2013/02/12" and the properties "gitsrc" and "gitbranch".
gitbranch can also be a tag or release.
The current version does not include any form of authentication towards the Git server.
ArtifactReferences can be used to include or exclude files.
If no include statement is found, everything is included by default, except explicitly excluded.
If a file is included as well as excluded, it will be excluded.

Signed-off-by: Armin Hueneburg <hueneburg.armin@gmail.com>