Add matching functionality
For each open requirement found in a topology a suitable candidate with the
matching capability is searched in the provider repositories.
Furthermore, the matching relationship between the requirement and capability is searched.
Not only single nodes but also topolgoy fragments can be injected.

Furthermore, the matching of hosts used for the splitting function is replaced by this
advanced matching function.

Modify the API to enable the injection of the topology fragments.

Signed-off-by: Karoline Saatkamp <karoline.saatkamp@iaas.uni-stuttgart.de>