Add missing Apache commons-codec to the yaml reader
Using the org.eclipse.winery.yaml.common module in another project
produces a error if the program is packaged with mvn package and
afterwards run from the command line. The error message tells that the
DigestUtils class laying in the Apache commons-codec package can not be
found.
This commit adds the commons-codec dependency to the org.eclipse.winery.yaml.common
modules pom.xml, which resolves the error.

Signed-off-by: Nick Fode <nickfode@posteo.de>