Extend Relationship Types to specify configure interface
This change introduces the possibility in order to define "Interfaces"
(in addition to Source and Target Interfaces) on Relationship Types.
The new Interface is basically the implementation according to the new
YAML specification. Here, one has the possibility to define a
"Configure" interface with a couple of predefined operations (see spec
for details).

Signed-off-by: Michael Wurster <miwurster@googlemail.com>