Fix Very Long CSAR Export Times
Integrity checks (syncrhonization between definitions and the files actually stored in the repository)
for ArtifactTemplates and SelfServiceMetadata were causing a long delay in CSAR export (2-3 minutes) due
to bad implementation (unnecessary writes to the repository were invoked). This is now fixed and export time
is in the range of 10 seconds.

Furthermore, a class that helps in debugging execution time issues was introduced.

Signed-off-by: Ghareeb Falazi <ghareeb.falazi@hotmail.com>