Enable FileBasedRepository imports through CsarImporter (#425)
The capability to import CSARs to a FileBasedRepository without git support is crucial to the introduction of the winery model into the container, cf. OpenTOSCA/container#78.

Notably for correct processing, validation and other purposes, CSARs within the container need to be segregated, which is only possible if the storage locations are different. Since the CsarImporter does not expose an overload that allows specifying the target repository, the RepositoryFactory needs to be reconfigured to take a specific FileBasedRepositoryConfiguration.

Signed-off-by: Clemens Lieb <vogel612@gmx.de>