Stop removing artifact-local references during import
The import process used to remove artifact references to local files stored inside the artifact.
This implies that during runtime the artifact references attached to a TArtifactTemplate obtained
from an IRepository are not necessarily the references that would be stored in the exported version.

That behaviour would require adjustments for the deployment process of a CSAR that duplicate the
code used during export and for unit testing within the winery (cf. BackendUtils#synchronizeReferences).
It's cleaner to allow the ArtifactReference information to be generally applicable instead of
working around the current behaviour.

Signed-off-by: Clemens Lieb <vogel612@gmx.de>