Remove Definitions in canonical model.
This change is useful because it simplifies the existing serialization mechanisms around the canonical model
and the original justification for Definitions having an anonymous type derived from tDefinitions is moot
for winery purposes. This is primarily because the winery REST API does not expose specific features
that rely on Definitions being extensible.

In addition this prepares the separation of XML and YAML providers and fully diverges from the schema definition
outlined in the XML-1.0 standard, which was no longer tenable when considering the incompatibilities between
XML and YAML standard..